some improvements in session managment

+++ decentraveller-app/context/AppContext.tsx
@@ -6,7 +6,9 @@ import AsyncStorage from '@react-native-async-storage/async-storage';
 
 export const AppContext = React.createContext<AppContextType | null>(null);
 
-const DEFAULT_CHAIN_ID = 31337;
+const DEFAULT_CHAIN_ID = 5;
+const DEFAULT_CHAIN_ID_HEX = '0x5';
+
 const UNRECOGNIZED_CHAIN_ID_MESSAGE = (chainId) =>
     `Unrecognized chain ID "${chainId}". Try adding the chain using wallet_addEthereumChain first.`;
 
@@ -17,6 +19,7 @@ interface UpdateSessionPayloadParams {
 
 const AppContextProvider: React.FC<React.ReactNode> = ({ children }) => {
     const [connectionContext, setConnectionContext] = React.useState<ConnectionContext>(null);
+    const connectionContextRef = React.useRef<ConnectionContext>(null);
     const [subscriptionsDone, setSubscriptionsDone] = React.useState<boolean>(false);
 
     const [nickname, setUserNickname] = React.useState<string>('');
@@ -27,15 +30,16 @@ const AppContextProvider: React.FC<React.ReactNode> = ({ children }) => {
     const connector = useWalletConnect();
 
     const pushChangeUpdate = async () => {
+        console.log('Trying to update chain...');
         try {
             await connector.sendCustomRequest({
                 id: 1,
                 jsonrpc: '2.0',
                 method: 'wallet_switchEthereumChain',
-                params: [{ chainId: '0x7a69' }],
+                params: [{ chainId: DEFAULT_CHAIN_ID_HEX }],
             });
         } catch (e) {
-            if (e.message === UNRECOGNIZED_CHAIN_ID_MESSAGE('0x7a69')) {
+            if (e.message === UNRECOGNIZED_CHAIN_ID_MESSAGE(DEFAULT_CHAIN_ID_HEX)) {
                 try {
                     await connector.sendCustomRequest({
                         id: 2,
@@ -43,7 +47,7 @@ const AppContextProvider: React.FC<React.ReactNode> = ({ children }) => {
                         method: 'wallet_addEthereumChain',
                         params: [
                             {
-                                chainId: '0x7a69',
+                                chainId: DEFAULT_CHAIN_ID_HEX,
                                 chainName: 'Local Hardhat',
                                 nativeCurrency: {
                                     nameText: 'Ethereum',
@@ -87,29 +91,39 @@ const AppContextProvider: React.FC<React.ReactNode> = ({ children }) => {
 
     const userLocation: AppContextStateArg<[string, string]> = memoFactory(location, setLocation);
 
-    /*React.useEffect(() => {
-         This effect allow us to clean sessions that the WalletConnect connector stores in the AsyncStorage
+    React.useEffect(() => {
+        // This effect allow us to clean sessions that the WalletConnect connector stores in the AsyncStorage
         const wipeAsyncStorage = async () => {
-            await AsyncStorage.clear();
-            setConnectionContext(null);
+            if (!connector.connected) {
+                await AsyncStorage.clear();
+                setConnectionContext(null);
+            }
         };
 
         wipeAsyncStorage().catch((e) => console.log('There was a problem wiping async storage', e));
-    }, []);*/
+    }, []);
 
     const updateConnectionContext = (address, chainId) => {
+        console.log('Updating connection context', address, chainId);
         const isWrongChain = chainId !== DEFAULT_CHAIN_ID;
-        setConnectionContext({
+        const newConnectionContext = {
             connectedAddress: address,
             connectedChainId: chainId,
             isWrongChain: isWrongChain,
-        });
+        };
+        setConnectionContext(newConnectionContext);
+        connectionContextRef.current = newConnectionContext;
+    };
+
+    const cleanConnectionContext = () => {
+        setConnectionContext(null);
+        connectionContextRef.current = null;
     };
 
     React.useEffect(() => {
-        console.log('connected:', connector.connected);
+        /*console.log('connected:', connector.connected);
         console.log('session:', connector.session);
-        console.log(connector.chainId);
+        console.log(connector.chainId);*/
 
         if (connector.connected) {
             updateConnectionContext(connector.accounts[0], connector.chainId);
@@ -126,7 +140,16 @@ const AppContextProvider: React.FC<React.ReactNode> = ({ children }) => {
                     const params: UpdateSessionPayloadParams = payload.params[0];
                     console.log(params);
                     console.log('session_update');
-                    if (connectionContext && params.chainId !== connectionContext.connectedChainId) {
+                    /*
+                     * The session_update event is not received when the chain is switched to a custom network, so
+                     * this callback does not work for changes to our local development blockchain. This has been
+                     * tested and only happens for our hardhat network. The problem with this is that there are
+                     * scenarios in which we ask the user to change the chain although the correct one is used.
+                     * */
+                    if (
+                        connectionContextRef.current &&
+                        params.chainId !== connectionContextRef.current.connectedChainId
+                    ) {
                         updateConnectionContext(params.accounts[0], params.chainId);
                     }
                 });
@@ -134,7 +157,7 @@ const AppContextProvider: React.FC<React.ReactNode> = ({ children }) => {
         }
     }, [connector.connected]);
 
-    const cleanConnectionContext = () => setConnectionContext(null);
+    console.log('cn context', connectionContext);
 
     return (
         <AppContext.Provider
